Add friends list feature with pagination and refactor home page structure
